@@ -1,26 +1,33 @@
 buildscript {
     repositories {
-        maven { url 'https://maven.fabric.io/public' }
+        google()
+        mavenCentral()
     }
 
     dependencies {
-        classpath 'io.fabric.tools:gradle:1.+'
     }
 }
+
+
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-android'
-apply plugin: 'io.fabric'
+
+def apikeyPropertiesFile = rootProject.file("apikey.properties");
+def apikeyProperties = new Properties()
+apikeyProperties.load(new FileInputStream(apikeyPropertiesFile))
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 32
     defaultConfig {
         applicationId "com.semyon.cucon"
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 32
         versionCode 1
         versionName "0.1.3 BETA"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        buildConfigField("String", "cryptocompare_key", apikeyProperties['cryptocompare_key'])
+        buildConfigField("String", "currencyconverterapi_key", apikeyProperties['currencyconverterapi_key'])
     }
     buildTypes {
         release {
@@ -32,16 +39,14 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:support-v4:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'com.google.firebase:firebase-core:16.0.7'
-    implementation 'com.crashlytics.sdk.android:crashlytics:2.9.8'
-    implementation 'com.google.firebase:firebase-core:16.0.7'
+    implementation 'com.google.firebase:firebase-core:20.0.2'
+    implementation 'com.google.firebase:firebase-core:20.0.2'
+    implementation 'com.google.android.material:material:1.4.0'
 }
 repositories {
     mavenCentral()
-    maven { url 'https://maven.fabric.io/public' }
+    google()
 }
+
 apply plugin: 'com.google.gms.google-services'
